refactor the bq client
